requirements.txt
asgiref==3.8.1
attrs==25.4.0
dj-database-url==2.3.0
dj-rest-auth==7.0.1
django==5.2.9
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.2
django-simple-history==3.10.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dotenv==0.9.9
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==24.1
pkgconfig==1.5.5
postgres==4.0
psycopg2-binary==2.9.11
psycopg2-pool==1.2
pyjwt==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pyyaml==6.0.3
referencing==0.37.0
rpds-py==0.28.0
six==1.17.0
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.1
uritemplate==4.2.0
whitenoise==6.8.2